Trying to get deal builder to work
I need help getting this to work.
-I mostly focused on the free deal component.
-Child component calls initial save from parent component method. If
the deal prop in child component is not null then assign the object an
id, check that fields are filled out the call save passing the deal
prop.
-In the save method of parent component, if deal parameter does not
have an id, create a new deal in db. If it does have id, then find if
changedDeals array contains a deal object with same id, if it does
splice, if not then concat.
-Once person clicks “Save” on parent component, send changedDeals array
to update db

——————————————————————
Issues:
-When I refresh my page data in free deals does not load until I
navigate away from the route and come back. The organization.js handler
doesn’t seem to be re rendering after updating the deals.

-When I do go to another route and come back, the data that is
displayed is the initially saved data, not the updated data. Wonder if
data is even updating in the db.