Deal Builder Better Saving
Before this commit, a deal would be inserted into the db as soon as you
blurred out of a field and all the fields in the deal were filled out.
Although it worked, it caused some unwanted rendering behavior before
user ever even pressed “save”. Now new deals are saved to an array, and
when the Save button is pressed, the array is sent to the api to be
saved.