Route Work, UX flow redesign, public profile showing
This is a working flow of how I think the users should be introduced to
our app.

Users will only be able to see a “manage” menu item if they have
created a business, which they can find a link to do so when traversing
deals.

I changed the Organizations menu item to Deals, and present users with
a categories page where they can find businesses based on category. I
changed the categories as well to be more encompassing of all business
types. The change in categories will mean that organizations you have
created previously, will have different categories assigned in their db
values, and thus will not show in the category page. You need to go in
and update the categories of you businesses in order for them to be
displayed in the category page.

Once they click on a category, they will find the list of businesses
with the associated category. When this is clicked, they will see the
businesses public profile.

Right now the UI is kinda all over the place, but the basic
functionality is there. UI updates to be done shortly.