Dynamic Categories, Locations Issue Resolved
Deal Categories now display only categories that actually contain
deals. Customers will never click into a category to find that no deals
are being offered there.

Locations had a bug that was not showing a newly added location after
navigating away to another route and coming back. I had to move the get
call to the actions up in the route hierarchy.

I added the header and side bar to the the category view as well as the
business profile view.