Submit to Review Functionality
On a managed organization dashboard, when all the info is in the right
spot, you will now see a button to submit the org to review. If you
click the button, it will set the org’s status to review and disable
all fields relating to org’s info.

Prop user’s first name was being passed as firstName instead of
first_name. Had to make appropriate changes.