Partial complete to be rebased
completed organizations_user_manages to display views for manager and
superadmin.

began to change the our logic from filtering all orgs to find if a
person is a manager to just having a db field for role in the user
table. (instead of an isSuperAdmin field).

Save changed from organizationStatus to status.