Added the Limited Redemptions Functionality
This shows the limit of deals left for a customer. When customer can
redeem the deal and the redemption will be recorded along. When you go
back to the deal, you will see that the limit has gone down. If it is
at 0 and you click on the deal, you will be forwarded to an
add_redemptions page.