Mailgun Promise
Made a change to mailgun send function to be a promise.
In password_reset, I changed the order of conditional statements for
clarity.
In Users Action, I wrapped a error respons in function that checked if
error actually exists. Without this, the error message would show to
the user for a split second before a successful return from the api.